requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.11
django-cors-headers==2.0.2
djangorestframework==3.6.2
gunicorn==19.7.1
mysqlclient==1.3.10
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
whitenoise==3.3.0
